refactor(plugin): extract limited/{escalate,video_source_timeline}.rs (#1092)
Begin splitting gpu_context/mod.rs's LimitedAccess vtable section into
banner-bounded submodules under limited/.

* limited/escalate.rs (~110 LoC) — Phase C3 escalate_begin/end token
  bridge. Mints + retires the scope tokens that every FullAccess
  callback validates via scope_token::with_full_scope_or_err.

* limited/video_source_timeline.rs (~190 LoC) — v12 (#958) set / clear
  / wait callbacks for the host's published per-frame
  HostVulkanTimelineSemaphore, plus the v14 (#1066) host_arc accessor
  the cdylib uses to construct a borrow without re-dispatching per
  frame.

Visibility on every callback is pub(in
crate::core::plugin::host_services) — narrowest scope that lets the
parent mod.rs's HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE static reach
the function pointers via the limited/ re-export.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
