+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/mod.rs
@@ -25,9 +25,15 @@ use super::host_callbacks;
 use super::run_host_extern_c;
 use super::shared::wire::{slice_from_raw, write_err, write_id_bytes};
 
+mod limited;
 mod scope_token;
 mod shared;
 
+use limited::{
+    host_gpu_lim_clear_video_source_timeline_semaphore, host_gpu_lim_escalate_begin,
+    host_gpu_lim_escalate_end, host_gpu_lim_host_video_source_timeline_arc,
+    host_gpu_lim_set_video_source_timeline_semaphore, host_gpu_lim_wait_timeline_semaphore,
+};
 use scope_token::with_full_scope_or_err;
 use shared::{handle_as_gpu_context, pixel_format_from_raw};
 
@@ -282,181 +288,6 @@ unsafe extern "C" fn host_gpu_lim_texture_native_dma_buf_fd(
             }
         },
         -1,
-    )
-}
-
-// -------------------------------------------------------------------------
-// Video-source timeline semaphore publish/clear (v12 — #958)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_set_video_source_timeline_semaphore(
-    handle: *const c_void,
-    timeline_handle: *const c_void,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_set_video_source_timeline_semaphore",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return;
-            };
-            if timeline_handle.is_null() {
-                return;
-            }
-            #[cfg(target_os = "linux")]
-            {
-                // SAFETY: `timeline_handle` is a borrowed
-                // `Arc::as_ptr(&Arc<HostVulkanTimelineSemaphore>)`
-                // produced by the cdylib caller. Bump the refcount so
-                // we can take a temporary owned Arc via `Arc::from_raw`;
-                // the caller's Arc strong-count is unchanged.
-                // Mirrors the `host_gpu_lim_register_texture` pattern
-                // for borrowed `Arc<TextureInner>`-shaped handles.
-                let ptr = timeline_handle
-                    as *const crate::vulkan::rhi::HostVulkanTimelineSemaphore;
-                unsafe {
-                    Arc::increment_strong_count(ptr);
-                }
-                let arc = unsafe { Arc::from_raw(ptr) };
-                gpu.set_video_source_timeline_semaphore(&arc);
-                // `arc` drops here, balancing the `increment_strong_count`
-                // above. The slot holds its own `Arc::clone` (taken by
-                // `set_video_source_timeline_semaphore` from the
-                // borrow).
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = timeline_handle;
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_clear_video_source_timeline_semaphore(
-    handle: *const c_void,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_clear_video_source_timeline_semaphore",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return;
-            };
-            #[cfg(target_os = "linux")]
-            {
-                gpu.clear_video_source_timeline_semaphore();
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = gpu;
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_wait_timeline_semaphore(
-    _handle: *const c_void,
-    timeline_handle: *const c_void,
-    value: u64,
-    timeout_ns: u64,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_wait_timeline_semaphore",
-        || {
-            // `gpu_handle` is intentionally ignored — the timeline
-            // borrow carries its own `vulkanalia::Device`, so the
-            // wait runs against the timeline directly without
-            // dereferencing any `GpuContext` instance. The handle
-            // stays in the wire format for cross-slot consistency.
-            if timeline_handle.is_null() {
-                write_err(
-                    "wait_timeline_semaphore: null timeline_handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            #[cfg(target_os = "linux")]
-            {
-                // SAFETY: `timeline_handle` is a borrowed pointer
-                // from the cdylib's
-                // `HostVulkanTimelineSemaphore::wait_via_vtable`
-                // (which gets it via `self as *const Self`). The
-                // host borrow lasts only for the duration of the
-                // wait call. We call `wait_direct` to bypass the
-                // `host_callbacks().is_some()` check on `wait()`
-                // itself — otherwise the host would re-dispatch
-                // through the vtable into infinite recursion.
-                let timeline = unsafe {
-                    &*(timeline_handle
-                        as *const crate::vulkan::rhi::HostVulkanTimelineSemaphore)
-                };
-                match timeline.wait_direct(value, timeout_ns) {
-                    Ok(()) => 0,
-                    Err(e) => {
-                        write_err(
-                            &format!("wait_timeline_semaphore: {e}"),
-                            err_buf,
-                            err_buf_cap,
-                            err_len,
-                        );
-                        1
-                    }
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = (timeline_handle, value, timeout_ns);
-                write_err(
-                    "wait_timeline_semaphore: Linux-only",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                1
-            }
-        },
-        1,
-    )
-}
-
-// -------------------------------------------------------------------------
-// host_video_source_timeline_arc — v14 (#1066)
-// -------------------------------------------------------------------------
-
-/// Clone the host's `Arc<HostVulkanTimelineSemaphore>` from the
-/// publish slot and return the raw `Arc::into_raw` pointer to the
-/// cdylib. The cdylib reconstitutes via `Arc::from_raw`; the host's
-/// slot retains its own independent strong count. Returns null when
-/// `gpu_handle` is null or when no producer has published a
-/// timeline (the slot is `None`).
-unsafe extern "C" fn host_gpu_lim_host_video_source_timeline_arc(
-    handle: *const c_void,
-) -> *const c_void {
-    run_host_extern_c(
-        "host_gpu_lim_host_video_source_timeline_arc",
-        || -> *const c_void {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return std::ptr::null();
-            };
-            #[cfg(target_os = "linux")]
-            {
-                match gpu.video_source_timeline_semaphore() {
-                    Some(arc) => Arc::into_raw(arc) as *const c_void,
-                    None => std::ptr::null(),
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = gpu;
-                std::ptr::null()
-            }
-        },
-        std::ptr::null(),
     )
 }
 
@@ -746,106 +577,6 @@ unsafe extern "C" fn host_gpu_lim_unregister_texture(
             gpu.unregister_texture(id_str);
         },
         (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// Escalate scope transition (Phase C3)
-// -------------------------------------------------------------------------
-
-/// Begin an escalate scope on the supplied `gpu_handle`. Mints a
-/// unique opaque token via
-/// [`crate::core::context::escalate_scope_registry::begin_escalate_scope`]
-/// and writes it into `*out_scope_token`. Blocking on the gate is
-/// expected — the host's escalate gate serializes against any
-/// concurrent escalate scope on the same `GpuContext`.
-unsafe extern "C" fn host_gpu_lim_escalate_begin(
-    handle: *const c_void,
-    out_scope_token: *mut *const c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_escalate_begin",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "escalate_begin: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1i32;
-            };
-            if out_scope_token.is_null() {
-                write_err(
-                    "escalate_begin: null out_scope_token",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1i32;
-            }
-            // begin_escalate_scope clones the Arc into the registry
-            // and enters the gate; both operations succeed without
-            // returning a fallible value.
-            let token = crate::core::context::escalate_scope_registry::begin_escalate_scope(
-                Arc::clone(gpu),
-            );
-            // SAFETY: out_scope_token is non-null per the check above.
-            // Token encoding is just the u64 serial reinterpreted as
-            // pointer-shaped; cdylib treats it as opaque.
-            unsafe { *out_scope_token = token as *const c_void };
-            0
-        },
-        1,
-    )
-}
-
-/// End an escalate scope. Removes the bound `Arc<GpuContext>` from
-/// the registry (releasing the escalate gate), then runs
-/// [`GpuContext::wait_device_idle`] to match the host-mode escalate
-/// path's scope-end semantics. Idempotent for stale or never-issued
-/// tokens.
-unsafe extern "C" fn host_gpu_lim_escalate_end(
-    _handle: *const c_void,
-    scope_token: *const c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_escalate_end",
-        || {
-            let token = scope_token as u64;
-            // Resolve the Arc BEFORE removing it from the registry so
-            // we can call wait_device_idle. If the token is stale or
-            // never-issued, this returns None — silently no-op (the
-            // gate was never acquired by this token, so there's
-            // nothing to release).
-            let arc_clone = crate::core::context::escalate_scope_registry::with_scope(
-                token,
-                Arc::clone,
-            );
-            let removed = crate::core::context::escalate_scope_registry::end_escalate_scope(token);
-            if !removed {
-                return 0i32;
-            }
-            match arc_clone.as_ref().map(|arc| arc.wait_device_idle()) {
-                Some(Ok(())) | None => 0,
-                Some(Err(e)) => {
-                    write_err(
-                        &format!("escalate_end: wait_device_idle failed: {e}"),
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    1
-                }
-            }
-        },
-        1,
     )
 }
